добавил шедулеры, и исправление багов

+++ db/models.py
@@ -14,7 +14,7 @@ class User(Base):
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
 
-    tg_id: Mapped[int] = mapped_column(Integer)
+    tg_id: Mapped[int] = mapped_column(BigInteger)
 
     username: Mapped[str] = mapped_column(String, nullable=True)
